Move functions to its respective files
