+++ src/transaction/deploy_account.rs
@@ -383,6 +383,51 @@ impl DeployAccount {
     }
 }
 
+// ----------------------------------
+//      Try from starknet api
+// ----------------------------------
+
+impl TryFrom<starknet_api::transaction::DeployAccountTransaction> for DeployAccount {
+    type Error = SyscallHandlerError;
+
+    fn try_from(
+        value: starknet_api::transaction::DeployAccountTransaction,
+    ) -> Result<Self, SyscallHandlerError> {
+        let max_fee = value.max_fee.0;
+        let version = Felt252::from_bytes_be(value.version.0.bytes());
+        let nonce = Felt252::from_bytes_be(value.nonce.0.bytes());
+        let class_hash: [u8; 32] = value.class_hash.0.bytes().try_into().unwrap();
+        let contract_address_salt = Felt252::from_bytes_be(value.contract_address_salt.0.bytes());
+
+        let signature = value
+            .signature
+            .0
+            .iter()
+            .map(|f| Felt252::from_bytes_be(f.bytes()))
+            .collect();
+        let constructor_calldata = value
+            .constructor_calldata
+            .0
+            .as_ref()
+            .iter()
+            .map(|f| Felt252::from_bytes_be(f.bytes()))
+            .collect();
+
+        let chain_id = Felt252::zero();
+
+        DeployAccount::new(
+            class_hash,
+            max_fee,
+            version,
+            nonce,
+            constructor_calldata,
+            signature,
+            contract_address_salt,
+            chain_id,
+        )
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use std::{path::PathBuf, sync::Arc};
