+++ src/transaction/invoke_function.rs
@@ -20,7 +20,7 @@ use crate::{
 use crate::services::api::contract_classes::deprecated_contract_class::EntryPointType;
 use cairo_vm::felt::Felt252;
 use getset::Getters;
-use num_traits::Zero;
+use num_traits::{One, Zero};
 
 use super::{fee::charge_fee, Transaction};
 
@@ -393,6 +393,97 @@ pub(crate) fn preprocess_invoke_function_fields(
                 Ok((entry_point_selector_field, additional_data))
             }
             None => Err(TransactionError::InvokeFunctionNonZeroMissingNonce),
+        }
+    }
+}
+
+// ----------------------------------
+//      Try from starknet api
+// ----------------------------------
+
+fn convert_invoke_v0(
+    value: starknet_api::transaction::InvokeTransactionV0,
+) -> Result<InvokeFunction, TransactionError> {
+    let contract_address = Address(Felt252::from_bytes_be(
+        value.contract_address.0.key().bytes(),
+    ));
+    let max_fee = value.max_fee.0;
+    let entry_point_selector = Felt252::from_bytes_be(value.entry_point_selector.0.bytes());
+    let version = Felt252::zero();
+    let nonce = None;
+    let chain_id = Felt252::zero();
+
+    let signature = value
+        .signature
+        .0
+        .iter()
+        .map(|f| Felt252::from_bytes_be(f.bytes()))
+        .collect();
+    let calldata = value
+        .calldata
+        .0
+        .as_ref()
+        .iter()
+        .map(|f| Felt252::from_bytes_be(f.bytes()))
+        .collect();
+
+    InvokeFunction::new(
+        contract_address,
+        entry_point_selector,
+        max_fee,
+        version,
+        calldata,
+        signature,
+        chain_id,
+        nonce,
+    )
+}
+
+fn convert_invoke_v1(
+    value: starknet_api::transaction::InvokeTransactionV1,
+) -> Result<InvokeFunction, TransactionError> {
+    let contract_address = Address(Felt252::from_bytes_be(value.sender_address.0.key().bytes()));
+    let max_fee = value.max_fee.0;
+    let version = Felt252::one();
+    let nonce = Felt252::from_bytes_be(value.nonce.0.bytes());
+    let chain_id = Felt252::zero();
+    let entry_point_selector = EXECUTE_ENTRY_POINT_SELECTOR.clone();
+
+    let signature = value
+        .signature
+        .0
+        .iter()
+        .map(|f| Felt252::from_bytes_be(f.bytes()))
+        .collect();
+    let calldata = value
+        .calldata
+        .0
+        .as_ref()
+        .iter()
+        .map(|f| Felt252::from_bytes_be(f.bytes()))
+        .collect();
+
+    InvokeFunction::new(
+        contract_address,
+        entry_point_selector,
+        max_fee,
+        version,
+        calldata,
+        signature,
+        chain_id,
+        Some(nonce),
+    )
+}
+
+impl TryFrom<starknet_api::transaction::InvokeTransaction> for InvokeFunction {
+    type Error = TransactionError;
+
+    fn try_from(
+        value: starknet_api::transaction::InvokeTransaction,
+    ) -> Result<Self, TransactionError> {
+        match value {
+            starknet_api::transaction::InvokeTransaction::V0(v0) => convert_invoke_v0(v0),
+            starknet_api::transaction::InvokeTransaction::V1(v1) => convert_invoke_v1(v1),
         }
     }
 }
